Added fat jar capability
