+++ build.gradle
@@ -29,14 +29,32 @@ dependencies {
 
     // Use JUnit test framework
     testImplementation 'junit:junit:4.12'
+
+    /* uncomment for cross-platform jar: */
+//    runtimeOnly "org.openjfx:javafx-graphics:$javafx.version:win"
+//    runtimeOnly "org.openjfx:javafx-graphics:$javafx.version:linux"
+//    runtimeOnly "org.openjfx:javafx-graphics:$javafx.version:mac"
 }
 
 application {
     // Define the main class for the application.
-    mainClassName = 'sample.Main'
+    mainClassName = 'sample.App'
 }
 
 javafx {
     version = "12"
-    modules = [ 'javafx.controls', 'javafx.fxml']
+    modules = ['javafx.controls', 'javafx.fxml']
+}
+
+mainClassName = 'sample.App'
+
+
+//for a fat jar
+jar {
+    manifest {
+        attributes 'Main-Class': 'sample.Main'
+    }
+    from {
+        configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) }
+    }
 }
